+++ javav2/example_code/sqs/src/main/java/com/example/sqs/HelloSQS.java
@@ -1,0 +1,49 @@
+//snippet-sourcedescription:[HelloSQS.java demonstrates how to  list queues by using a paginated response.]
+//snippet-keyword:[AWS SDK for Java v2]
+//snippet-keyword:[Amazon Simple Notification Service]
+/*
+   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+   SPDX-License-Identifier: Apache-2.0
+*/
+
+package com.example.sqs;
+// snippet-start:[sqs.java2.hello.main]
+import software.amazon.awssdk.auth.credentials.EnvironmentVariableCredentialsProvider;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.sqs.SqsClient;
+import software.amazon.awssdk.services.sqs.model.SqsException;
+import software.amazon.awssdk.services.sqs.paginators.ListQueuesIterable;
+
+/**
+ * Before running this Java V2 code example, set up your development environment, including your credentials.
+ *
+ * For more information, see the following documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+ */
+public class HelloSQS {
+
+    public static void main(String[] args) {
+        SqsClient sqsClient = SqsClient.builder()
+            .region(Region.US_WEST_2)
+            .credentialsProvider(EnvironmentVariableCredentialsProvider.create())
+            .build();
+
+        listQueues(sqsClient);
+        sqsClient.close();
+    }
+
+    public static void listQueues(SqsClient sqsClient) {
+        try {
+            ListQueuesIterable listQueues = sqsClient.listQueuesPaginator();
+            listQueues.stream()
+                .flatMap(r -> r.queueUrls().stream())
+                .forEach(content -> System.out.println(" Queue URL: " + content.toLowerCase()));
+
+        } catch (SqsException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+}
+// snippet-end:[sqs.java2.hello.main]
