+++ javav2/example_code/sqs/src/main/resources/config.properties
@@ -1,3 +1,5 @@
-QueueName = <enter queue name>
-DLQueueName =<enter a dl queue name>
-Message= <enter a message>
+QueueName = MyTestQueue31
+DLQueueName = MyDLQueue
+Message= Test message sent by the SQS Service
+kmsMasterKeyAlias = aws/sqs
+tagName = Accounting ID
